+++ lib/models/sub_category/sub_category.dart
@@ -1,5 +1,5 @@
+import 'package:four_jars/models/main_category_type/main_category_type.dart';
 import 'package:hive/hive.dart';
-import 'package:four_jars/models/main_category_type.dart';
 
 part 'sub_category.g.dart';
