+++ lib/screens/category_details/category_details_controller.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:four_jars/logic/budget_manager.dart';
+import 'package:four_jars/models/transaction/transaction.dart';
 
-import 'package:four_jars/models/transaction.dart';
 import 'package:four_jars/screens/home/widgets/add_transaction_sheet/add_transaction_sheet.dart';
 import 'package:four_jars/screens/home/widgets/add_transaction_sheet/add_transaction_sheet_controller.dart';
 import 'package:provider/provider.dart';
