+++ lib/screens/settings/settings_controller.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:four_jars/logic/budget_manager.dart';
-import 'package:four_jars/models/main_category_type.dart';
+import 'package:four_jars/models/main_category_type/main_category_type.dart';
 import 'package:hive/hive.dart';
 
 class SettingsController extends ChangeNotifier {
