+++ lib/screens/settings/settings_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:four_jars/models/main_category_type.dart';
+import 'package:four_jars/models/main_category_type/main_category_type.dart';
 import 'package:four_jars/screens/settings/settings_controller.dart';
 import 'package:four_jars/screens/settings/sub_category_list/sub_category_list_screen.dart';
 import 'package:four_jars/theme/theme_controller.dart';
